Restructure source tree of server
