+++ main.py
@@ -1,102 +1,10 @@
 from __future__ import annotations
 
-import asyncio
-import logging
-from collections import OrderedDict
-from contextlib import AbstractAsyncContextManager
-from types import TracebackType
-from typing import Final, Optional, Type, TYPE_CHECKING
-
-from fastapi import FastAPI
-from fastapi.responses import RedirectResponse
-
-try:
-    import coverage  # dev-dependency only
-except ImportError:
-    pass
-
-try:
-    import uvloop  # type: ignore
-except ImportError:
-    pass
-else:
-    asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
-
-from server import api_v1, CI
+from server import global_app
 
 
-logger = logging.getLogger(__name__)
-subapps = OrderedDict()
-subapps["/api/v1"] = api_v1
-
-final_subroute = list(subapps.keys())[-1]
-final_subapp = list(subapps.values())[-1]
+__all__ = ("app",)
 
 
-class ApplicationLifespan(AbstractAsyncContextManager):
-
-    __slots__ = ("app", "cov")
-    if TYPE_CHECKING:
-        app: Final[FastAPI]
-        cov: Optional[coverage.Coverage]
-
-    def __init__(self, app: FastAPI) -> None:
-        self.app = app
-
-        self.cov = None
-        if CI:
-            self.cov = coverage.Coverage(data_suffix=True, config_file=True)
-            self.cov.start()
-            logger.info("Measuring code coverage...")
-
-    async def __aenter__(self) -> None:
-        coros = [a.router.lifespan_context(subapp).__aenter__() for a in subapps.values()]
-        await asyncio.gather(*coros)
-
-    async def __aexit__(
-        self,
-        exc_type: Optional[Type[BaseException]],
-        exc_val: Optional[BaseException],
-        exc_tb: Optional[TracebackType],
-    ) -> None:
-        if self.cov is not None:
-            self.cov.stop()
-            self.cov.save()
-
-        coros = [a.router.lifespan_context(subapp).__aexit__(exc_type, exc_val, exc_tb) for a in subapps.values()]
-        await asyncio.gather(*coros)
-
-
-app = FastAPI(
-    title="Resident manager API",
-    docs_url=None,
-    redoc_url=None,
-    lifespan=ApplicationLifespan,
-    version=final_subapp.version,
-)
-for route, subapp in subapps.items():
-    app.mount(route, subapp)
-
-
-@app.get("/", include_in_schema=False)
-async def root() -> RedirectResponse:
-    """Redirect to API documentation of latest version"""
-    return RedirectResponse(final_subroute)
-
-
-@app.get("/loop", include_in_schema=False)
-async def loop() -> str:
-    """Return current asyncio event loop"""
-    return str(asyncio.get_event_loop())
-
-
-@app.get("/docs", include_in_schema=False)
-async def docs() -> RedirectResponse:
-    """Redirect to API documentation of latest version"""
-    return RedirectResponse(final_subroute + "/docs")
-
-
-@app.get("/redoc", include_in_schema=False)
-async def redoc() -> RedirectResponse:
-    """Redirect to API documentation of latest version"""
-    return RedirectResponse(final_subroute + "/redoc")
+# Expose application to uvicorn
+app = global_app
