+++ server/__init__.py
@@ -1,2 +1,3 @@
 from .config import *
+from .globals import *
 from .v1 import *
