+++ server/v1/index.html
@@ -4,6 +4,7 @@
 <head>
     <meta charset="utf-8"> <!-- Important: rapi-doc uses utf8 characters -->
     <title>Resident manager API v1</title>
+    <link rel="icon" type="image/x-icon" href="/api/v1/static/favicon.png">
     <script type="module" src="https://unpkg.com/rapidoc/dist/rapidoc-min.js"></script>
 </head>
