+++ server/v1/routes/__init__.py
@@ -1,1 +1,4 @@
-from .api import *
+from .admin import *
+from .login import *
+from .register import *
+from .residents import *
