+++ server/v1/routes/admin/registration_requests/accept.py
@@ -4,17 +4,17 @@ from typing import List, Optional
 
 from fastapi import Response, status
 
-from .....app import api_v1
-from .....models import AuthorizationHeader, Resident, Result, Snowflake
+from ....app import api_v1
+from ....models import AuthorizationHeader, RegisterRequest, Result, Snowflake
 
 
-__all__ = ("admin_residents_delete",)
+__all__ = ("admin_reg_request_accept",)
 
 
 @api_v1.post(
-    "/admin/residents/delete",
-    name="Account deletion",
-    description="Delete one or more resident accounts",
+    "/admin/registration-requests/accept",
+    name="Registration requests approval",
+    description="Approve one or more registration requests",
     tags=["admin"],
     response_model=None,
     responses={
@@ -28,7 +28,7 @@ __all__ = ("admin_residents_delete",)
     },
     status_code=status.HTTP_204_NO_CONTENT,
 )
-async def admin_residents_delete(
+async def admin_reg_request_accept(
     headers: AuthorizationHeader,
     response: Response,
     objects: List[Snowflake],
@@ -38,5 +38,5 @@ async def admin_residents_delete(
         response.status_code = status.HTTP_400_BAD_REQUEST
         return auth
 
-    await Resident.delete_many(objects)
+    await RegisterRequest.accept_many(objects)
     return None
