+++ server/v1/routes/admin/registration_requests/reject.py
@@ -4,8 +4,8 @@ from typing import List, Optional
 
 from fastapi import Response, status
 
-from .....app import api_v1
-from .....models import AuthorizationHeader, RegisterRequest, Result, Snowflake
+from ....app import api_v1
+from ....models import AuthorizationHeader, RegisterRequest, Result, Snowflake
 
 
 __all__ = ("admin_reg_request_reject",)
