+++ server/v1/routes/admin/residents/__init__.py
