+++ server/v1/routes/admin/residents/delete.py
@@ -4,17 +4,17 @@ from typing import List, Optional
 
 from fastapi import Response, status
 
-from .....app import api_v1
-from .....models import AuthorizationHeader, Result, RoomData
+from ....app import api_v1
+from ....models import AuthorizationHeader, Resident, Result, Snowflake
 
 
-__all__ = ("admin_rooms_delete",)
+__all__ = ("admin_residents_delete",)
 
 
 @api_v1.post(
-    "/admin/rooms/delete",
-    name="Room information deletion",
-    description="Delete room information",
+    "/admin/residents/delete",
+    name="Account deletion",
+    description="Delete one or more resident accounts",
     tags=["admin"],
     response_model=None,
     responses={
@@ -28,15 +28,15 @@ __all__ = ("admin_rooms_delete",)
     },
     status_code=status.HTTP_204_NO_CONTENT,
 )
-async def admin_rooms_delete(
+async def admin_residents_delete(
     headers: AuthorizationHeader,
     response: Response,
-    rooms: List[int],
+    objects: List[Snowflake],
 ) -> Optional[Result[None]]:
     auth = await headers.verify_admin()
     if auth is not None:
         response.status_code = status.HTTP_400_BAD_REQUEST
         return auth
 
-    await RoomData.delete_many(rooms)
+    await Resident.delete_many(objects)
     return None
