+++ server/v1/routes/admin/rooms/__init__.py
