+++ server/v1/routes/admin/rooms/count.py
@@ -4,21 +4,21 @@ from typing import Optional
 
 from fastapi import Response, status
 
-from .....app import api_v1
-from .....models import AuthorizationHeader, Resident, Result
+from ....app import api_v1
+from ....models import AuthorizationHeader, Result, Room
 
 
-__all__ = ("admin_residents_count",)
+__all__ = ("admin_rooms_count",)
 
 
 @api_v1.get(
-    "/admin/residents/count",
-    name="Residents count",
-    description="Return number of residents",
+    "/admin/rooms/count",
+    name="Rooms count",
+    description="Return number of rooms",
     tags=["admin"],
     responses={
         status.HTTP_200_OK: {
-            "description": "Return number of residents",
+            "description": "Return number of rooms",
             "model": Result[int],
         },
         status.HTTP_400_BAD_REQUEST: {
@@ -27,17 +27,15 @@ __all__ = ("admin_residents_count",)
         },
     },
 )
-async def admin_residents_count(
+async def admin_rooms_count(
     headers: AuthorizationHeader,
     response: Response,
-    id: Optional[int] = None,
-    name: Optional[str] = None,
     room: Optional[int] = None,
-    username: Optional[str] = None,
+    floor: Optional[int] = None,
 ) -> Result[Optional[int]]:
     auth = await headers.verify_admin()
     if auth is not None:
         response.status_code = status.HTTP_400_BAD_REQUEST
         return auth
 
-    return Result(data=await Resident.count(id=id, name=name, room=room, username=username))
+    return Result(data=await Room.count(room=room, floor=floor))
