+++ server/v1/routes/admin/rooms/root.py
@@ -4,9 +4,9 @@ from typing import List, Optional
 
 from fastapi import Response, status
 
-from .....app import api_v1
-from .....models import AuthorizationHeader, Result, Room
-from ......config import DB_PAGINATION_QUERY
+from ....app import api_v1
+from ....models import AuthorizationHeader, Result, Room
+from .....config import DB_PAGINATION_QUERY
 
 
 __all__ = ("admin_rooms",)
