--- server/v1/routes/api/__init__.py
@@ -1,4 +1,0 @@
-from .admin import *
-from .login import *
-from .register import *
-from .residents import *
