+++ server/v1/routes/register.py
@@ -4,8 +4,8 @@ from typing import Annotated, Optional
 
 from fastapi import Query, Response, status
 
-from ...app import api_v1
-from ...models import (
+from ..app import api_v1
+from ..models import (
     AuthorizationHeader,
     PersonalInfo,
     PublicInfo,
