+++ server/v1/routes/residents/__init__.py
