+++ server/v1/tests/__init__.py
@@ -1,0 +1,1 @@
+from .test_main import *
